refactor: 재고 수량에 대한 데이터 타입 통일

+++ product-service/src/main/java/com/example/productservice/jpa/ProductStockEntity.java
@@ -20,31 +20,12 @@ public class ProductStockEntity {
     private Long productId;
 
     @Column(name = "stock")
-    private Long stock;
+    private Integer stock;
 
-    public ProductStockEntity of(Long productId, Long stock) {
+    public ProductStockEntity of(Long productId, Integer stock) {
         ProductStockEntity productStockEntity = new ProductStockEntity();
         productStockEntity.setProductId(productId);
         productStockEntity.setStock(stock);
         return productStockEntity;
     }
-
-//     // 재고 증가
-//    public void increaseStock(Long quantity) {
-//        if (quantity < 0) {
-//            throw new VitaQueueException(ErrorCode.STOCK_INCREASE_NEGATIVE, "재고 증가 수량은 음수일 수 없습니다.");
-//        }
-//        this.stock += quantity;
-//    }
-//
-//    // 재고 감소
-//    public void decreaseStock(Long quantity) {
-//        if (quantity < 0) {
-//            throw new VitaQueueException(ErrorCode.STOCK_DECREASE_NEGATIVE, "재고 감소 수량은 음수일 수 없습니다.");
-//        }
-//        if (this.stock < quantity) {
-//            throw new VitaQueueException(ErrorCode.STOCK_NOT_ENOUGH, "재고가 부족합니다.");
-//        }
-//        this.stock -= quantity;
-//    }
 }
